Refactor: 채널요청 API
