+++ src/constants/revalidate-tag.ts
@@ -1,10 +1,19 @@
+// application
 export const METADATAS_TAG = 'metadatas';
+
+// channel & schedule
 export const CHANNELS_TAG = 'channels';
-export const CHANNEL_COUNT_TAG = 'channelCount';
-export const WAITING_TAG = 'waiting';
-export const FEATURED_TAG = 'featured';
 export const SCHEDULES_TAG = 'schedules';
+
+// personal
 export const MEMBER_TAG = 'member'; // 단건
-export const SETLISTS_TAG = 'setlists'; // 다건
 export const BLACKLIST_TAG = 'blacklist';
 export const WHITELIST_TAG = 'whitelist';
+
+// special
+export const SETLISTS_TAG = 'setlists'; // 다건
+export const FEATURED_TAG = 'featured';
+
+// request-channel
+export const CHANNEL_COUNT_TAG = 'channelCount';
+export const WAITING_TAG = 'waiting';
